% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/afs_forecast.R
\name{AFSForecast}
\alias{AFSForecast}
\title{This function outputs a time
series afs forecast for Leases, Credit Cards, and other
variables, from time series input files for bank position
data, macroeconomic forecasts , and arima model
coefficients.}
\usage{
AFSForecast(position_data, model_coefficients, macro_forecasts)
}
\arguments{
\item{position_data}{A data frame of a particular bank's present balance sheet}

\item{model_coefficients}{A matrix of arima calibration coefficients computed from past history}

\item{macro_forecasts}{An MTS of either basic, adverse, or severely adverse macroeconomic forecasts}
}
\value{
object of class MTS
}
\description{
This function outputs a time
series afs forecast for Leases, Credit Cards, and other
variables, from time series input files for bank position
data, macroeconomic forecasts , and arima model
coefficients.
}
\author{
David Karapetyan
}

