% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/balance_forecast.R
\name{BalanceForecast}
\alias{BalanceForecast}
\title{This function outputs a time
series balance forecast from a time series input file for
bank position data.}
\usage{
BalanceForecast(position_data, macro_forecasts)
}
\arguments{
\item{position_data}{A data frame of a particular
bank's present balance sheet}

\item{macro_forecasts}{A time series of either basic, adverse, or severely
adverse macroeconomic forecasts}
}
\value{
object of class MTS
}
\description{
This function outputs a time
series balance forecast from a time series input file for
bank position data.
}
\author{
David Karapetyan
}

