% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prepare_position_data.R
\name{PreparePositionData}
\alias{PreparePositionData}
\title{This function creates position data for a bank for a given quarter.}
\usage{
PreparePositionData(bank, quarter, nco_data, ppnr_data, total_assets,
  capital_data)
}
\arguments{
\item{bank}{A character string that exists in each of the data files.}

\item{quarter}{A character string in the format "2013Q1".}

\item{nco_data}{A data frame object aggregating U.S. bank NCO data}

\item{ppnr_data}{A data frame object aggregating U.S. bank NCO data}

\item{total_assets}{A data  Total assets data in data.frame format.}

\item{capital_data}{Capital data in data.frame format.}
}
\value{
A data frame of one row that contains all the merged data.
}
\description{
This function creates position data for a bank for a given quarter.
}

