% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prepare_for_plot.R
\name{PrepareForPlot}
\alias{PrepareForPlot}
\title{This function outputs an object suitable for plotting in GGplot.}
\usage{
PrepareForPlot(book, variable, bank, quarter, nco_data, ppnr_data, total_assets,
  capital_data, model_coefficients, macro_forecasts)
}
\arguments{
\item{book}{Portion of trading book from which variables are obtained. This
can be any of the following \code{c("lll", "afs", "loss", "nco", "capital",
"balance", "ppnr")}}

\item{variable}{Variable that will be plotted, with respect to time}

\item{model_coefficients}{A data frame of arima calibration
coefficients computed from past history}

\item{macro_forecasts}{A time series of either basic, adverse, or severely
adverse macroeconomic forecasts}

\item{position_data}{A data frame of a particular
bank's present balance sheet}
}
\value{
Object of class data.frame
}
\description{
This function outputs an object suitable for plotting in GGplot.
}
\author{
David Karapetyan
}

